pyproject.toml
[project]
name = "EDM"
version = "0.0.1"
authors = [
    {name="WizardSuey", email="dorogino@gmail.com"},
]
description = "EDM"
readme = "README.md"
dependencies = [
    "flask",
    "flask-socketio",
    "flask-script",
    "psycopg2-binary",
    "Flask-WTF",
    "flask-ckeditor",
]
license = "MIT"
license-files = ["LICENSE"]

[project.urls]
Homepage = "https://github.com/WizardSuey"

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"
